Inline code snippets from the samples
This PR adds a transfrom that inlines the follwing code snippet reference from the samples repo:

```
<!-- atomist:code-snippet:start=lib/sdm/dotnetCore.ts#dotnetGenerator -->
Anything can be in here
<!-- atomist:code-snippet:end-->
```

Once this is placed inside a markdown file in the docs repo, it will replace the section with the contents from the `lib/sdm/dotnetCore.ts`. See https://github.com/atomist/samples/blob/master/lib/sdm/dotnetCore.ts#L36-L53 for how the sections are marked in the samples.